fix build warning
/home/mi/xiaomi/test/cNet_IO/client.c:56:16: error: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   56 |         if(num == strlen(sendbuf)+1)
      |                ^~
/home/mi/xiaomi/test/cNet_IO/client.c:61:9: error: implicit declaration of function ‘sleep’ [-Werror=implicit-function-declaration]
   61 |         sleep(1);
      |         ^~~~~
/home/mi/xiaomi/test/cNet_IO/client.c:64:5: error: implicit declaration of function ‘close’; did you mean ‘pclose’? [-Werror=implicit-function-declaration]
   64 |     close(cliSocket);
      |     ^~~~~
      |     pclose

+++ client.c
@@ -5,6 +5,7 @@
 #include <netinet/in.h>
 #include <arpa/inet.h>
 #include <string.h>
+#include <unistd.h>
 
 #define DEFAULT_IP        "127.0.0.1"
 #define DEFAULT_PORT    6666
@@ -53,7 +54,7 @@ int main(int argc, char *argv[])
     {
 
         int num = send(cliSocket, sendbuf, strlen(sendbuf)+1, 0);
-        if(num == strlen(sendbuf)+1)
+        if(num == (int)strlen(sendbuf)+1)
             printf("Send: Hello Server, I am Client.\nlength:%d\n\n",num);
         else
             printf("send err\n");
